requirements.txt
Flask
pandas-datareader
yfinance
requests
pandas
prophet
matplotlib
python-dotenv
Flask-Migrate
SQLAlchemy
Flask-WTF
werkzeug
flask_jwt_extended